Cloud Computing/requirements.txt
Flask
mysql-connector-python
bcrypt
pyjwt
google-cloud-storage

